backend/requirements.txt
google-generativeai
vosk
websockets
uvicorn[standard]
databases[asyncpg]
fastapi
# numpy
# soundfile
gTTS
